Adding files related to Back-propagation algorithm
To get started, you will need to download the starter
code and unzip its contents to the directory where you wish to complete the
exercise. If needed, use the cd command in Octave/MATLAB to change to
this directory before starting this exercise.
There is a pdf document included in the zipped files which contains instructions to execute the code.